# public/api-doc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Makkelijke Markt API",
        "version": "1.1.0"
    },
    "paths": {
...
        "/api/1.1.0/factuur/concept/{dagvergunningId}": {
            "post": {
                "tags": [
                    "Factuur"
                ],
                "summary": "Stelt factuur voor dagvergunning op zonder opslag",
                "operationId": "FactuurPostConcept",
                "parameters": [
                    {
                        "name": "dagvergunningId",
                        "in": "path",
                        "description": "ID van de dagvergunning",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Factuur"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [],
                        "bearer": []
                    }
                ]
            }
        },
        "/api/1.1.0/report/factuur/overzicht/{van}/{tot}": {
            "get": {
                "tags": [
                    "Factuur"
                ],
                "summary": "Geeft factuuren in periode vanaf <YYYY-mm-dd> tot <YYYY-mm-dd>",
                "operationId": "FactuurGetByRange",
                "parameters": [
                    {
                        "name": "van",
                        "in": "path",
                        "description": "als YYYY-mm-dd",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tot",
                        "in": "path",
                        "description": "als YYYY-mm-dd",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Factuur"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [],
                        "bearer": []
                    }
                ]
            }
        },
        "/api/1.1.0/report/factuur/overzichtmarkt/{marktId}/{van}/{tot}": {
            "get": {
                "tags": [
                    "Factuur"
                ],
                "summary": "Geeft facturen voor Markt in periode vanaf <YYYY-mm-dd> tot <YYYY-mm-dd>",
                "operationId": "FactuurGetByMarktAndRange",
                "parameters": [
                    {
                        "name": "marktId",
                        "in": "path",
                        "description": "ID van de markt",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "van",
                        "in": "path",
                        "description": "als YYYY-mm-dd",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tot",
                        "in": "path",
                        "description": "als YYYY-mm-dd",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [],
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Factuur": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "14"
                    },
                    "totaal": {
                        "type": "string",
                        "example": "10.01"
                    },
                    "exclusief": {
                        "type": "string",
                        "example": "12.11"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "MmAppKey",
                "in": "header"
            },
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}